+++ src/cdk/create-rest-api.ts
@@ -11,7 +11,7 @@ import {
   aws_route53_targets,
 } from 'aws-cdk-lib';
 import type {StackConfig} from '../get-stack-config';
-import {getDomainName} from '../utils/get-absolute-domain-name';
+import {getDomainName} from '../utils/get-domain-name';
 import {getHash} from '../utils/get-hash';
 import {getNormalizedName} from '../utils/get-normalized-name';
