+++ src/cdk/create-stack.ts
@@ -1,6 +1,6 @@
 import {App, Stack} from 'aws-cdk-lib';
 import type {StackConfig} from '../get-stack-config';
-import {getDomainName} from '../utils/get-absolute-domain-name';
+import {getDomainName} from '../utils/get-domain-name';
 import {getStackName} from '../utils/get-stack-name';
 
 export function createStack(stackConfig: StackConfig): Stack {
