+++ src/sdk/find-stack.ts
@@ -1,6 +1,6 @@
 import type {Stack} from '@aws-sdk/client-cloudformation';
 import type {StackConfig} from '../get-stack-config';
-import {getDomainName} from '../utils/get-absolute-domain-name';
+import {getDomainName} from '../utils/get-domain-name';
 import {getStackName} from '../utils/get-stack-name';
 import {findStacks} from './find-stacks';
